Move geolocation to its own file
